Use filter to apply wpautop
wpautop is a tough thing to always like. I found that since 2.3 my HTML emails with custom content templates have been broken. Removing autop from the filter fixed everything so I would love to have the option to remove the function by means of a filter because it's just a nuisance. Examples in my emails:

<code><a href="http://mywebsite.com">A Link</a></code>
would turn into:
<code><a<br />
href="http://mywebsite.com">A Link</a></code>

Same with the IMG tag and really any tag that accepted parameters. Not sure if this is just me, but I couldn't find a way around it except modifying my WooCommerce install.