Update cart-shipping.php
Would it be possible to add the $package as a condition for this filter? The majority of cases where I modify this message is when I need to change it based on cart statistics such as subtotal. Obtaining the cart details is simple enough, but then I had a client splitting the cart into different shipping packages, and each package alone has to reach a specific minimum, but I couldn't find a way to retrieve the cost for the given package without modifying WooCommerce, and modifying the template files defeats the purpose of my side code in a way. I'm trying to avoid copying this file just for this message, this way it is always compatible with future versions.

Appreciate it!